requirements.txt
django>=4.2
django-htmx>=1.17
pytest-django>=4.7
pytest-factoryboy>=2.7
python-dotenv>=1.0
